- Added type annotations to .map() operations - Added .toDouble() to constructors - Added exports to main library

+++ lib/src/audio_file.dart
@@ -24,14 +24,14 @@ class AudioFile {
     createdAt = (map['createdAt'] == null) ? null : DateTime.parse(map['createdAt']),
     createdBy = map['createdBy'],
     description = map['description'],
-    duration = map['duration'] ?? 0,
+    duration = (map['duration'] ?? 0).toDouble(),
     fileUrl = map['fileUrl'],
     id = map['id'] ?? 0,
-    voteAverage = map['voteAverage'] ?? 0,
+    voteAverage = (map['voteAverage'] ?? 0).toDouble(),
     voteCount = map['voteCount'] ?? 0,
-    voteWeightedAverage = map['voteWeightedAverage'] ?? 0,
+    voteWeightedAverage = (map['voteWeightedAverage'] ?? 0).toDouble(),
     word = map['word'];
-  
+
   factory AudioFile.fromJson(String jsonString) {
     return AudioFile.fromMap(json.decode(jsonString));
   }
